:children_crossing: Make minor adjustments to game
Bugs will now emit particles when being squashed. The particles will
relocate when the bug relocates.

The tutorial screen has been adjusted to visibly show the bug so that
players understand that they are looking for a visual bug, and not a
code bug.

This version of the game is now licensed under the MPLv2.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>